Do floating point math

+++ src/math_parser_diag.cpp
@@ -1338,7 +1338,7 @@ std::function<double( dialogue & )> calories_eval( char scope,
 
         if( format == "percent" ) {
             if( d.actor( beta )->get_character() ) {
-                int divisor = d.actor( beta )->get_healthy_kcal() / 100;
+                double divisor = d.actor( beta )->get_healthy_kcal() / 100.0;
                 //if no data, default to default height of 175cm
                 if( divisor == 0 ) {
                     divisor = 118169 / 100;
